+++ src/Blockchain/custom/Update-AzBlockchainMember.ps1
@@ -141,12 +141,12 @@ param(
 process {
     try {
         if ($PSBoundParameters.ContainsKey('Password')) {
-            $psTxt = [System.Runtime.InteropServices.marshal]::PtrToStringAuto([System.Runtime.InteropServices.marshal]::SecureStringToBSTR($PSBoundParameters['Password']))
+            $psTxt = . "$PSScriptRoot/../utils/Unprotect-SecureString.ps1" $PSBoundParameters['Password']
             $PSBoundParameters.Remove('Password')
             $PSBoundParameters.Add('Password', $psTxt)
         }
         if ($PSBoundParameters.ContainsKey('ConsortiumManagementAccountPassword')) {
-            $psTxt = [System.Runtime.InteropServices.marshal]::PtrToStringAuto([System.Runtime.InteropServices.marshal]::SecureStringToBSTR($PSBoundParameters['ConsortiumManagementAccountPassword']))
+            $psTxt = . "$PSScriptRoot/../utils/Unprotect-SecureString.ps1" $PSBoundParameters['ConsortiumManagementAccountPassword']
             $PSBoundParameters.Remove('ConsortiumManagementAccountPassword')
             $PSBoundParameters.Add('ConsortiumManagementAccountPassword', $psTxt)
         }