+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,7 +19,11 @@
 -->
 ## Upcoming Release
 
+## Version 1.9.5
+* Fixed DateTime parse issue in common libraries [#13045]
+
 ## Version 1.9.4
+* Formatted the upcoming breaking change messages
 * Updated Azure.Core to 1.4.1
 
 ## Version 1.9.3